src/referal_service_back/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "app as referal service for testing"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
